+++ src/pages/ProductList/styles.ts
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const ProductsWrapper = styled.div`
+  display: flex;
+  justify-content: start;
+  flex-wrap: wrap;
+  gap: 20px;
+  padding: 50px 240px;
+`;
+
+export { ProductsWrapper };
